Don't need to convert to int from Integer for AppdefEntityID

+++ src/org/hyperic/hq/ui/action/resource/common/monitor/visibility/ViewChartAction.java
@@ -6,7 +6,7 @@
  * normal use of the program, and does *not* fall under the heading of
  * "derived work".
  * 
- * Copyright (C) [2004, 2005, 2006, 2007], Hyperic, Inc.
+ * Copyright (C) [2004-2008], Hyperic, Inc.
  * This file is part of HQ.
  * 
  * HQ is free software; you can redistribute it and/or modify
@@ -77,7 +77,7 @@ public class ViewChartAction extends MetricDisplayRangeAction {
 
         AppdefEntityID adeId =
             new AppdefEntityID( chartForm.getType().intValue(),
-                                chartForm.getRid().intValue() );
+                                chartForm.getRid());
 
         HashMap forwardParams = new HashMap(3);
         forwardParams.put( Constants.RESOURCE_PARAM, chartForm.getRid() );
